asns/asn2/pyproject.toml
[project]
name = "asn2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.1.0",
    "pandas>=2.3.3",
]
